add movie-nav connect
